requirements.txt
python-telegram-bot~=13.7
tweepy~=3.10.0